+++ src/main.cpp
@@ -2,10 +2,6 @@
 
 #include <Geode/modify/GauntletSelectLayer.hpp>
 class $modify(GPGauntletSelectLayer, GauntletSelectLayer) {
-    struct Fields {
-        CCMenuItemSpriteExtra* m_timeButton;
-    };
-
     bool init(int idk) {
         if (!GauntletSelectLayer::init(idk)) return false;
 
@@ -13,50 +9,47 @@ class $modify(GPGauntletSelectLayer, GauntletSelectLayer) {
         auto timeMenu = CCMenu::create();
         timeMenu->setPosition(winSize.width - 30.0f, winSize.height - 30.0f);
         timeMenu->setID("time-menu"_spr);
+        timeMenu->setVisible(GameLevelManager::sharedState()->m_savedGauntlets->count() > 0);
         addChild(timeMenu, 2);
 
-        auto f = m_fields.self();
-        f->m_timeButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_timeIcon_001.png", 1.0f, [this](auto) {
-            GPGauntletsPopup::create([this] {
+        auto timeButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_timeIcon_001.png", 1.0f, [this](CCMenuItemSpriteExtra* sender) {
+            GPGauntletsPopup::create([this, sender] {
                 if (m_scrollLayer) {
                     m_scrollLayer->removeFromParent();
                     m_scrollLayer = nullptr;
                 }
                 m_leftButton->setVisible(false);
                 m_rightButton->setVisible(false);
+                sender->setVisible(false);
                 auto bottomRightMenu = getChildByID("bottom-right-menu");
                 if (bottomRightMenu) for (auto child : CCArrayExt<CCNode*>(bottomRightMenu->getChildren())) {
-                    if (child != m_refreshButton) child->removeFromParent();
+                    if (child != m_refreshButton) child->removeFromParent(); // gdutils
                 }
                 setupGauntlets();
                 m_leftButton->setVisible(true);
                 m_rightButton->setVisible(true);
+                sender->setVisible(true);
             })->show();
         });
-        f->m_timeButton->setID("time-button"_spr);
-        timeMenu->addChild(f->m_timeButton);
+        timeButton->setID("time-button"_spr);
+        timeMenu->addChild(timeButton);
 
         return true;
     }
 
     void setupGauntlets() {
-        auto f = m_fields.self();
-        if (f->m_timeButton) f->m_timeButton->setEnabled(false);
-
-        auto pageNav = getChildByID("dasshu.gauntletredesign/page-navigation");
-        if (pageNav) pageNav->removeFromParent();
-
         auto glm = GameLevelManager::sharedState();
-        std::vector<bool> projectedIDs;
+        auto projectedIDs = Mod::get()->getSavedValue("projected-ids", std::vector<bool>(NUM_GAUNTLETS, false));
+        projectedIDs.resize(NUM_GAUNTLETS, false);
         auto newSavedGauntlets = CCDictionary::create();
         newSavedGauntlets->retain();
         std::vector<bool> gauntletsAdded;
-        for (int i = 1; i < 53; i++) {
-            auto id = std::to_string(i);
-            if (!glm->m_savedGauntlets->objectForKey(id) && GPGauntletsPopup::projectedIDs[i - 1]) {
+        for (int i = 0; i < NUM_GAUNTLETS; i++) {
+            auto id = std::to_string(i + 1);
+            if (!glm->m_savedGauntlets->objectForKey(id) && projectedIDs[i]) {
                 gauntletsAdded.push_back(true);
                 auto gauntlet = GJMapPack::create();
-                gauntlet->m_packID = i;
+                gauntlet->m_packID = i + 1;
                 gauntlet->m_levels = CCArray::create();
                 gauntlet->m_levels->retain();
                 glm->m_savedGauntlets->setObject(gauntlet, id);
@@ -70,16 +63,15 @@ class $modify(GPGauntletSelectLayer, GauntletSelectLayer) {
 
         GauntletSelectLayer::setupGauntlets();
 
-        for (int i = 0; i < 52; i++) {
+        for (int i = 0; i < NUM_GAUNTLETS; i++) {
             if (gauntletsAdded[i]) glm->m_savedGauntlets->removeObjectForKey(std::to_string(i + 1));
         }
 
-        if (f->m_timeButton) f->m_timeButton->setEnabled(true);
+        if (auto timeMenu = getChildByID("time-menu"_spr)) timeMenu->setVisible(glm->m_savedGauntlets->count() > 0);
     }
 
     void onPlay(CCObject* sender) {
-        auto glm = GameLevelManager::sharedState();
-        if (glm->m_savedGauntlets->objectForKey(std::to_string(sender->getTag())))
+        if (GameLevelManager::sharedState()->m_savedGauntlets->objectForKey(std::to_string(sender->getTag())))
             GauntletSelectLayer::onPlay(sender);
     }
 };