+++ changelog.md
@@ -1,6 +1,7 @@
 # Gauntlet Projection Changelog
 ## v1.0.1 (2024-09-26)
-- Fixed an incompatibility with the mod "Gauntlet Touchup" by jaydasshu
+- Added saving and loading of selected gauntlets
+- Tweaked the layout of the projection popup
 
 ## v1.0.0 (2024-09-26)
 - Initial release