+++ changelog.md
@@ -1,4 +1,7 @@
 # Gauntlet Projection Changelog
+## v1.0.2 (2024-10-05)
+- Fixed GDUtils particle info button
+
 ## v1.0.1 (2024-09-26)
 - Added saving and loading of selected gauntlets
 - Tweaked the layout of the projection popup